feature:move-app-context-into-app
