+++ src/apps/app/index.ts
@@ -1,3 +1,4 @@
 export * from './app-context-config.interface.js';
 export * from './app-error.element.js';
 export * from './app.element.js';
+export * from './app.context.js';
