--- src/shared/context/index.ts
@@ -1,1 +1,0 @@
-export * from './app.context.js';
