+++ tsconfig.json
@@ -19,6 +19,9 @@
 		"baseUrl": ".",
 		"incremental": true,
 		"paths": {
+			// APPS
+			"@umbraco-cms/backoffice/app": ["src/apps/app"],
+
 			"@umbraco-cms/backoffice/external/lit": ["src/external/lit"],
 			"@umbraco-cms/backoffice/external/lodash": ["src/external/lodash"],
 			"@umbraco-cms/backoffice/external/monaco-editor": ["src/external/monaco-editor"],
@@ -40,7 +43,6 @@
 
 			// SHARED
 			"@umbraco-cms/backoffice/auth": ["src/shared/auth"],
-			"@umbraco-cms/backoffice/context": ["src/shared/context"],
 			"@umbraco-cms/backoffice/events": ["src/shared/umb-events"],
 			"@umbraco-cms/backoffice/icon": ["src/shared/icon-registry"],
 			"@umbraco-cms/backoffice/models": ["src/shared/models"],
